requirements.txt
Django>=5.0
python-dotenv>=1.0
spotipy>=2.23
pandas>=2.2
openpyxl>=3.1
APScheduler>=3.10
